# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telegram Music API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/users/{user_id}/profile": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "Profile",
        "operationId": "profile_v1_users__user_id__profile_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/profile": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "My Profile",
        "operationId": "my_profile_v1_me_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/users/{user_id}/follow": {
      "put": {
        "tags": [
          "social"
        ],
        "summary": "Follow",
        "operationId": "follow_v1_users__user_id__follow_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "social"
        ],
        "summary": "Unfollow",
        "operationId": "unfollow_v1_users__user_id__follow_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{user_id}/followers": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "Followers",
        "operationId": "followers_v1_users__user_id__followers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionOut"
                  },
                  "title": "Response Followers V1 Users  User Id  Followers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{user_id}/following": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "Following",
        "operationId": "following_v1_users__user_id__following_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionOut"
                  },
                  "title": "Response Following V1 Users  User Id  Following Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/social/feed": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "Friends Feed",
        "description": "What the people you follow have been listening to. Empty until you follow someone.",
        "operationId": "friends_feed_v1_social_feed_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendActivityOut"
                  },
                  "title": "Response Friends Feed V1 Social Feed Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/wrapped": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "Wrapped",
        "operationId": "wrapped_v1_me_wrapped_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2020
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/public-profile": {
      "put": {
        "tags": [
          "social"
        ],
        "summary": "Set Public Profile",
        "operationId": "set_public_profile_v1_me_public_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/me/notifications": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "Get Notification Prefs",
        "operationId": "get_notification_prefs_v1_me_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPrefsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "social"
        ],
        "summary": "Set Notification Prefs",
        "operationId": "set_notification_prefs_v1_me_notifications_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPrefsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPrefsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistCountOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "plays": {
            "type": "integer",
            "title": "Plays"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "plays"
        ],
        "title": "ArtistCountOut"
      },
...
      "ArtistRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "primary",
              "feature",
              "composer",
              "lyricist"
            ],
            "title": "Role"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "ArtistRef"
      },
...
      "BusiestDayOut": {
        "properties": {
          "day": {
            "type": "string",
            "title": "Day"
          },
          "plays": {
            "type": "integer",
            "title": "Plays"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "day",
          "plays"
        ],
        "title": "BusiestDayOut"
      },
...
      "ConnectionOut": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "is_pro": {
            "type": "boolean",
            "title": "Is Pro"
          },
          "is_following": {
            "type": "boolean",
            "title": "Is Following"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "first_name",
          "is_pro",
          "is_following"
        ],
        "title": "ConnectionOut"
      },
...
      "FollowOut": {
        "properties": {
          "following": {
            "type": "boolean",
            "title": "Following"
          },
          "changed": {
            "type": "boolean",
            "title": "Changed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "following",
          "changed"
        ],
        "title": "FollowOut"
      },
      "FriendActivityOut": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "played_at": {
            "type": "string",
            "format": "date-time",
            "title": "Played At"
          },
          "track": {
            "$ref": "#/components/schemas/TrackOut"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "first_name",
          "played_at",
          "track"
        ],
        "title": "FriendActivityOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationPrefsIn": {
        "properties": {
          "prefs": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Prefs"
          },
          "tz_offset_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 840.0,
                "minimum": -720.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Tz Offset Minutes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "NotificationPrefsIn"
      },
      "NotificationPrefsOut": {
        "properties": {
          "prefs": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Prefs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "prefs"
        ],
        "title": "NotificationPrefsOut"
      },
...
      "ProfileOut": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "is_pro": {
            "type": "boolean",
            "title": "Is Pro"
          },
          "followers": {
            "type": "integer",
            "title": "Followers"
          },
          "following": {
            "type": "integer",
            "title": "Following"
          },
          "playlists": {
            "type": "integer",
            "title": "Playlists"
          },
          "tracks_played": {
            "type": "integer",
            "title": "Tracks Played"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time",
            "title": "Joined At"
          },
          "is_me": {
            "type": "boolean",
            "title": "Is Me"
          },
          "is_following": {
            "type": "boolean",
            "title": "Is Following"
          },
          "top_artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistCountOut"
            },
            "type": "array",
            "title": "Top Artists"
          },
          "top_tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "type": "array",
            "title": "Top Tracks"
          },
          "public_playlists": {
            "items": {
              "$ref": "#/components/schemas/PublicPlaylistOut"
            },
            "type": "array",
            "title": "Public Playlists"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "first_name",
          "is_pro",
          "followers",
          "following",
          "playlists",
          "tracks_played",
          "joined_at",
          "is_me",
          "is_following",
          "top_artists",
          "top_tracks",
          "public_playlists"
        ],
        "title": "ProfileOut"
      },
...
      "PublicPlaylistOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tracks_count": {
            "type": "integer",
            "title": "Tracks Count"
          },
          "share_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Slug"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "tracks_count"
        ],
        "title": "PublicPlaylistOut"
      },
      "PublicProfileIn": {
        "properties": {
          "public": {
            "type": "boolean",
            "title": "Public"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "public"
        ],
        "title": "PublicProfileIn"
      },
...
      "TrackChannelRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "is_featured": {
            "type": "boolean",
            "title": "Is Featured"
          },
          "subscribers_count": {
            "type": "integer",
            "title": "Subscribers Count"
          },
          "joined": {
            "type": "boolean",
            "title": "Joined"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "username",
          "title",
          "is_featured",
          "subscribers_count",
          "joined"
        ],
        "title": "TrackChannelRef",
        "description": "The channel a track is credited to (ADR-003 §2-5).\n\nThe point of showing it is to send members to the channel that published the\nmusic, so the one picked is the one the viewer has *not* joined yet."
      },
      "TrackOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistRef"
            },
            "type": "array",
            "title": "Artists"
          },
          "album": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "has_thumb": {
            "type": "boolean",
            "title": "Has Thumb"
          },
          "channels_count": {
            "type": "integer",
            "title": "Channels Count"
          },
          "playable": {
            "type": "boolean",
            "title": "Playable"
          },
          "liked": {
            "type": "boolean",
            "title": "Liked",
            "default": false
          },
          "channel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrackChannelRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "palette": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Palette"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "artists",
          "album",
          "duration",
          "language",
          "year",
          "has_thumb",
          "channels_count",
          "playable"
        ],
        "title": "TrackOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WrappedOut": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "plays": {
            "type": "integer",
            "title": "Plays"
          },
          "minutes": {
            "type": "integer",
            "title": "Minutes"
          },
          "unique_tracks": {
            "type": "integer",
            "title": "Unique Tracks"
          },
          "active_days": {
            "type": "integer",
            "title": "Active Days"
          },
          "top_artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistCountOut"
            },
            "type": "array",
            "title": "Top Artists"
          },
          "top_tracks": {
            "items": {
              "$ref": "#/components/schemas/WrappedTrackOut"
            },
            "type": "array",
            "title": "Top Tracks"
          },
          "busiest_day": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BusiestDayOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "type": "array",
            "title": "Tracks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "year",
          "plays",
          "minutes",
          "unique_tracks",
          "active_days",
          "top_artists",
          "top_tracks"
        ],
        "title": "WrappedOut"
      },
      "WrappedTrackOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "plays": {
            "type": "integer",
            "title": "Plays"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "plays"
        ],
        "title": "WrappedTrackOut"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}